Complete refactor:
- implement a OOP structure with MetricProvider objects that return prometheus Metrics from the json log result

- implement an agnostic NvmeCollector that stores MetricProviders instances and delegates handling the details of Collect and Describe to them

- move NvmeCollector instantiation in a convenience function in the main package
